requirements.txt
aiosignal==1.3.1
aniso8601==9.0.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
dnspython==2.3.0
email-validator==1.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
exceptiongroup==1.1.1
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
Flask-WTF==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
modeltranslation==0.25
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.3
packaging==23.1
pathy==0.10.1
pluggy==1.0.0
preshed==3.0.8
pydantic==1.10.7
PyJWT==2.6.0
pytest-asyncio==0.21.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
selenium==3.141.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.12
srsly==2.4.6
thinc==8.1.10
tomli==2.0.1
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
wasabi==1.1.1
websocket-client==0.59.0
Werkzeug==2.3.4
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0
